+++ src/mobius/Stream.cpp
@@ -486,8 +486,8 @@ PUBLIC int Stream::getPitchTweak(int tweak)
 
 PUBLIC long Stream::deltaFrames(float* start, float* end)
 {
-	long bytes = (long)((long)end - (long)start);
-	long samples = bytes / sizeof(float);
+	ptrdiff_t bytes = end - start;
+	long samples = (long) bytes / sizeof(float);
 	long frames = samples / channels;
 	return frames;
 }
